Created double_updates_methods module and moved two functions there from weights module
