+++ natstats/weights.py
@@ -2,7 +2,7 @@
 
 import pandas as pd
 
-from natstats.helpers import reindex_and_fill, flip, _get_end_year
+from natstats.helpers import reindex_and_fill, flip
 from natstats._validation import _handle_axis
 
 
@@ -51,21 +51,4 @@ def reindex_to_update_periods(weights):
     ]
     
     return pd.concat(to_concat)
-
-
-def jan_adjust_weights(weights, direction='back'):
-    """Adjust Feb weights by one month so that weights start in Jan."""
-    if direction == 'back':
-        return weights.tshift(-1, freq='MS')
-    elif direction == 'forward':
-        return weights.tshift(1, freq='MS')
-
-
-def adjust_pre_doublelink(weights, start_year='2017', direction='back'): 
-    """Jan adjusts only the weights up to the end year."""
-    # Double update (Jan & Feb) starts in 2017
-    return pd.concat([
-        jan_adjust_weights(weights[:_get_end_year(start_year)], direction),
-        weights[start_year:],
-    ])
